token/requirements.txt
requests==2.26.0
python-dotenv==0.20.1
